Llms: removed and spread out llm.routes
